Create rollercoaster code
This code checks the eligibility of a person for a rollercoaster ride. If the person is eligible foe the ride it checks the age of the person and assigns prices of the tickets according to the age.If the person wants to click a photo it charges extra along with the ticket price.